chore(setup): Added UI for Android launcher and main application in Android platform, and added Kotlin/Wasm support
This commit adds the UI for the Android platform's launcher and main application, as well as Kotlin/Wasm support to the project. The Android code includes a simple MainActivity that uses the Compose library to set up the app layout, an ic_launcher_background drawable, and an ic_launcher_foreground vector drawable for the app icon. In addition, a JVMPlatform class is added to represent the Java platform, and the Constants file contains the SERVER_PORT constant. Finally, the build.gradle.kts file has been updated with new dependencies for Kotlin/Wasm, Ktor server, and AndroidX libraries.